backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
anthropic>=0.51.0
pandas>=2.2.0
python-dotenv>=1.0.0
